// SqGoods/ClientApp/src/api-client.generated.ts
export interface CategoryCreateModel {
    name: string;
    order: number;
    topOrder?: number | null;
}

export interface CategoryUpdateModel extends CategoryCreateModel {
    id: string;
}

export interface CategoryListModel extends CategoryUpdateModel {
    attributesCount: number;
    productCount: number;
}

export interface AttributeListModel {
    id: string;
    orderInCategory?: number | null;
    name: string;
    type: SgAttributeType;
    unit?: string | null;
    mandatory?: boolean | null;
    numberOfCategories: number;
    numberOfItems?: number | null;
}

export enum SgAttributeType {
    Boolean = 0,
    Integer = 1,
    Select = 2,
    SubSet = 3,
}
// ...
export interface AttributeSwapFormModel {
    categoryId: string;
    swaps: AttributeSwapFormItemModel[];
}

export interface AttributeSwapFormItemModel {
    originalId: string;
    currentId: string;
}